+++ spec/extended_email_reply_parser_spec.rb
@@ -5,7 +5,9 @@ describe ExtendedEmailReplyParser do
     expect(ExtendedEmailReplyParser::VERSION).not_to be nil
   end
 
-  it 'does something useful' do
-    expect(false).to eq(true)
+  describe "#read" do
+    subject { ExtendedEmailReplyParser.read('spec/email_fixtures/email_de_1.txt') }
+
+    it { is_expected.to be_kind_of Mail::Message }
   end
 end