+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## master (unreleased)
 ### Security
 ### Added
+- Convenience method `ExtendedEmailReplyParser.read "/path/to/email.eml"  # => Mail::Message`
+
 ### Deprecated
 ### Removed
 ### Fixed